backend/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
google-generativeai
slack-sdk
flask
flask-cors